fix: 타입 수정

+++ src/main/java/com/ODG/ODG_back/dto/participant/response/ParticipantListResponseDto.java
@@ -3,12 +3,14 @@ package com.ODG.ODG_back.dto.participant.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ParticipantListResponseDto {
     private Long participantId;
     private String name;
     private String transportType;
-    private Double lat;
-    private Double lng;
+    private BigDecimal lat;
+    private BigDecimal lng;
 }
